backend/requirements.txt
# Core
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Database
pymongo[srv]>=4.5.0

# LangChain / LangGraph
langchain>=1.0.0
langchain-groq>=1.0.0
langchain-community>=0.4.0
langgraph>=1.0.0
langgraph-checkpoint>=3.0.0

# NLP
spacy>=3.7.0
textblob>=0.17.1
nltk>=3.8.0
textstat>=0.7.0
yake>=0.4.8
keybert>=0.8.0
sentence-transformers>=2.2.0

# LlamaIndex
llama-index-core>=0.14.0
llama-cpp-python>=0.2.90

# Web Scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
trafilatura>=1.6.0
ddgs>=1.0.0

# Utilities
numpy>=1.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
